Adição da mecânica de pontuação e tempo e modificação nas fases
Adicionado nas fases relógios que funcionam como pontuação + tempo que exibe no final do jogo + modificações das fases com partes mais difícies

+++ Assets/Scripts/Saw.cs
@@ -210,7 +210,7 @@ public class Saw : MonoBehaviour
         affectedByTime = false;
     }
 
-    public void SecondsToBeAffectedByTime(int seconds)
+    public void SecondsToBeAffectedByTime(float seconds)
     {
         Invoke("BeAffectedByTime", seconds);
     }
